dtor

Module declarative

Declarative forms of the #[dtor] macro.

The declarative forms wrap and parse a proc_macro-like syntax like so, and are identical in expansion to the undecorated procedural macros. The declarative forms support the same attribute parameters as the procedural macros.

dtor::declarative::dtor! {
  #[dtor]
  fn foo() {
    libc_println!("Goodbye, world!");
  }
}

// ... the above is identical to:

#[dtor]
fn foo() {
  libc_println!("Goodbye, world!");
}

Macros

dtor

Parse a #[dtor]-annotated item as if it were a proc-macro.